Enhancement: Extract IsNullTest
