+++ tests/Constraint/ConstraintTest.php
@@ -1142,56 +1142,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsNull()
-    {
-        $constraint = Assert::isNull();
-
-        $this->assertFalse($constraint->evaluate(0, '', true));
-        $this->assertTrue($constraint->evaluate(null, '', true));
-        $this->assertEquals('is null', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate(0);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that 0 is null.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintIsNull2()
-    {
-        $constraint = Assert::isNull();
-
-        try {
-            $constraint->evaluate(0, 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that 0 is null.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintIsNotNull()
     {
         $constraint = Assert::logicalNot(
